Update prepare_belle2.sh
this should reflect the current build requirements for the externals.
One open issue is tcl/tk: They are needed to build gitk, git-gui and
python.TkInter which is used for the interactive matplotlib.

I put them into optional because it works but we should discuss whether
gitk and git-gui are important enough to always build. I'm not really
worried about matplotlib but as far as I am aware  python.TkInter cannot
be installed by pip afterwards so users need to recompile the externals
to get interactive plotting support