Add support to setup externals without any release
this adds a new function/alias setupext which takes one or no argument
and just sets up a given externals version or the highest installed
version if no version was supplied.

It piggybacks on the normal update_environment() but just supplies None
for central and local release. update_environment() got a new optional
parameter to specify the externals version directly.