Add default SConscript when creating analysis directory using newana
When users use newmod to create a new module they need a SConscript file
to link against framework and possibly they want to link against
analysis and mdst_dataobjects as well.

So we should create a default SConscript which contains the most useful
libraries so that newmod works out of the box without them having to
copy one. If the analysis is only using python this does not hurt and
since the linker will remove libraries it doesn't need automatically
this should be close to optimal.

(Part of BII-1828)