Prepare to switch to pycodestyle (BII-5098)
autopep8 aleady used pycodestyle behind the scenes so we get
inconsistent results with pep8 due to new and improved warnings.

With this PR we disable all these extra warnings to get the same
behavior with pep8 and pycodestyle. The new externals will then remove
pep8 and provide a symlink pep8->pycodestyle so the tools don't need any
additional change for now.