Add some simple test infrastructure (BII-7037)
This commits adds a few scripts and instructions to test the tools on
different distributions using docker.

This is already setup in Bamboo to run test/run.sh in all supported
distributions to check simple setup for both bash and tcsh

It also adds man to the requirements as thisroot.csh executes man and
thus sourcing would fail in a `tcsh -e` that exits on errors as needed
for the test.