Make sure relative sourcing of b2setup.sh works
we only need the path to b2anypython in the very beginning before we
setup the PATH. Afterwards we can just call it.

So we can get rid of the extra variable which might pollute the shell
and this also fixes sourcing e.g. `../tools/b2setup.sh` and then
changing the directory: it will just find b2anypython in the PATH